backend/app/config.py
import os
from pathlib import Path
from dotenv import load_dotenv

# Load .env from the backend folder explicitly
env_path = Path(__file__).resolve().parent.parent / ".env"
load_dotenv(dotenv_path=env_path)

class Settings:
    # Database
    DB_HOST: str = os.getenv("DB_HOST", "127.0.0.1")
    DB_PORT: int = int(os.getenv("DB_PORT", 3306))
    DB_NAME: str = os.getenv("DB_NAME", "resume_db")
    DB_USER: str = os.getenv("DB_USER", "resume_user")
    DB_PASSWORD: str = os.getenv("DB_PASSWORD", "aditya1411")

    # CORS
    CORS_ORIGINS: str = os.getenv("CORS_ORIGINS", "http://localhost:5173")

    # Groq API
    GROQ_API_KEY: str = os.getenv("GROQ_API_KEY", "")
    GROQ_MODEL: str = os.getenv("GROQ_MODEL", "llama3-8b-8192")

    @property
    def DATABASE_URL(self) -> str:
        return "sqlite:///./test.db"


# Instantiate settings
settings = Settings()
