backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
SQLAlchemy==2.0.32
pymysql==1.1.1
alembic==1.13.2
python-dotenv==1.0.1
spacy==3.7.5
yake==0.4.8
sentence-transformers==3.0.1
PyPDF2==3.0.1
python-docx==1.1.2
groq==0.9.0

